+++ internal/driver/gomobile/canvas.go
@@ -137,7 +137,7 @@ func (c *mobileCanvas) Overlays() fyne.OverlayStack {
 }
 
 func (c *mobileCanvas) PixelCoordinateForPosition(pos fyne.Position) (int, int) {
-	return int(float32(pos.X) * c.scale), int(float32(pos.Y) * c.scale)
+	return int(pos.X * c.scale), int(pos.Y * c.scale)
 }
 
 func (c *mobileCanvas) Refresh(obj fyne.CanvasObject) {
